+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+- [#69] Fix: Base64URL valid characters are not properly validated
+- [#67] Add SECURITY.md
+
 ## [0.1.0-beta.9] - 2024-09-03
 
 - [#64] maintain iOS 15 compatibility